Restore parts of initial semantics by integrating the rebalance listener
